src/third_party/blink/web_tests/platform/linux/external/wpt/offscreen-canvas/text/2d.text.measure.width.basic-expected.txt

This is a testharness.js-based test.
FAIL The width of character is same as font used for OffscreenCanvas assert_equals: ctx.measureText('A').width === 50 (got 36[number], expected 50[number]) expected 50 but got 36
Harness: the test ran to completion.
